move the tools out of the main
