+++ corsika_iact_caller/__init__.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python
 """
 Call CORSIKA with the IACT package in a thread safe way
 
@@ -20,64 +19,23 @@ Notes:
     Returns the corsika return code.
 
 """
-from __future__ import absolute_import, print_function, division
-
-__all__ = ['corsika_iact']
-
 import docopt
+from .tools import all_files_in
+from .tools import mkdir
+from .tools import rm_dir
+from .tools import symlink
+from .tools import Path
+from .tools import supposed_to_store_output_path
+from .tools import extract_output_path_from
+from .tools import CanNotFindTelfilKeyWordInCorsikaInputCard
+from .tools import output_path_taken_from_corsika_input_card
+from .tools import extract_output_path_from
+
 import os
 import glob
 import subprocess
 import sys
 import random
-
-def all_files_in(path):
-    return glob.glob(os.path.join(path, '*'))
-
-def mkdir(path):
-    subprocess.call(['mkdir', path])
-
-def rm_dir(path):
-    subprocess.call(['rm', '-r', path])
-
-def symlink(src, dest):
-    subprocess.call(['ln', '-s', src, dest])
-
-class Path:
-    def __init__(self, path):
-        self.path = path
-        self.basename = os.path.basename(self.path)
-        self.basename_wo_extension = os.path.splitext(self.basename)[0]
-        self.dirname = os.path.dirname(self.path)
-
-def supposed_to_store_output_path(line):
-    if line[0:6] == 'TELFIL':
-        return True
-    else:
-        return False
-
-def extract_output_path_from(line):
-    first_quote = line.find('"')
-    second_quote = line[first_quote+1:].find('"') + first_quote+1
-    return line[first_quote+1:second_quote]
-
-class CanNotFindTelfilKeyWordInCorsikaInputCard(Exception):
-    pass
-
-def output_path_taken_from_corsika_input_card(input_card_path):
-    output_path = ""
-    found_TELFIL_keyword = False
-    with open(input_card_path) as fileobject:
-        for line in fileobject:
-            if supposed_to_store_output_path(line):
-                found_TELFIL_keyword = True
-                output_path = extract_output_path_from(line)
-
-    if found_TELFIL_keyword:
-        return output_path
-    else:
-        raise CanNotFindTelfilKeyWordInCorsikaInputCard
-
 
 def corsika_iact(corsika_path, input_card_path, save_stdout=False):
     """
